requirements.txt
numpy
pillow
tensorflow
scipy